requirements.txt
opencv-python
numpy
spotipy
python-dotenv
deepface
